Revert "Fixed casting"
This reverts commit c35f9a955087843ab6ddc7dda91ecc28362b2fa7.

+++ source/Cosmos.HAL2/Audio/AudioBuffer.cs
@@ -133,7 +133,7 @@ namespace Cosmos.HAL.Audio {
                 MemoryOperations.Copy(
                     dest,
                     bufferPtr + bufferOffset,
-                    (uint)channelByteSize
+                    channelByteSize
                 );
             }
         }
